Enable reuse of low-level netlink functions
Factors out basic netlink functions into the nlmsg package.
