+++ netlink/nlmsg/nlmsg_linux_test.go
@@ -1,0 +1,3 @@
+package nlmsg
+
+import ()
