+++ netlink/nlmsg/nlmsg_unsupported.go
@@ -1,0 +1,11 @@
+// +build !linux
+
+package nlmsg
+
+import (
+	"errors"
+)
+
+var (
+	ErrNotImplemented = errors.New("not implemented")
+)
